backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
flask-sqlalchemy==3.1.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
pymysql==1.1.0
bcrypt==4.0.1
flask-cors==4.0.0
python-jose==3.3.0
requests==2.31.0
openai==1.3.7
httpx>=0.24.1
mysql-connector-python==8.2.0